Switched PWM to 1ms implementation; protect shared state with mutex
PWM was refactored to use a struct to store state, and that struct was shared with the timer functions.
So mutex the struct since it is a shared resource, and switch to a 1ms PWM implementation because
stopping a timer blocked on a mutex results in UB.

Ideally we would store curTimeMs inside the state instead of a static
local, but casting away const could mess with aliasing guarantees.